Emit PrePack information as global into bundle (optionally)
Summary:
Emit PrePack info into the generated bundle. The intention is for this to include information about enabled prepack options. Currently it only includes a flag if lazy objects are enabled.

Its emitted as global which is somewhat intrusive so its guarded by a flag currently and defaults to false.

Example:
```
var __PREPACK__ = {lazyObjects: false}
```

Reviewed By: NTillmann

Differential Revision: D8321765

fbshipit-source-id: 03a41b2cef84ea44c02c71e233b407e3ddc1197d