Don't purge bindings when execution leaves scope
Release notes: None

For historical and apparently no longer needed reason,
we used to forget about modified bindings when execution
leaves scopes.

However, this means that the binding no longer properly participates
in state tracking.

While by itself, it didn't seem to cause much (or any?) harm,
it turned out to be a blocker for #2125, where the initial state
of a binding becomes important when joining partially leaked bindings.