Link to Prepack introduction from website (#2147)
Summary:
I haven't written the next parts yet but I figured let's put a link on the website so this doesn't get lost.
I think even the details there are useful enough for users and new contributors.

<img width="1513" alt="screen shot 2018-06-21 at 3 57 52 pm" src="https://user-images.githubusercontent.com/810438/41727325-f9ba32da-756b-11e8-94f0-c83b71df859b.png">

I put it at the top of Getting Started because IMO you'll want to read this (or at least see it) before trying to run Prepack.

I initially planned to put it in the docs themselves but only today realized there's no build process. I don't want to convert the guide to HTML so I'd rather keep it as a gist (and update it if necessary) in the meantime.
Closes https://github.com/facebook/prepack/pull/2147

Differential Revision: D8577403

Pulled By: gaearon

fbshipit-source-id: 2f7f64a0a4413066739b12c50b4d21779a16a568