Slice argument list to fix abstract recovery bug (#2156)
Summary:
Release notes: none

Fixes a bug where the argument list gets mutated and padded with undefined values, so instead we slice it a copy of it so we don't mutate the same array and pass the mutated array instead. Fixes https://github.com/facebook/prepack/issues/2153.
Closes https://github.com/facebook/prepack/pull/2156

Differential Revision: D8610746

Pulled By: trueadm

fbshipit-source-id: 1b6f37693323c813d60f43ab7e9977a99f86ff47