Fix require/define detection (#2168)
Summary:
Release Notes: None

It is okay for a define call to happen within a require call so long as the define call has happened by the time the require call has returned (since all define calls should have the same effect and only happen once). Previously if we encountered a module required before it was defined (or defined in a nested partial evaluation scope), we would not record it in ModuleTracer logic.
Closes https://github.com/facebook/prepack/pull/2168

Differential Revision: D8688788

Pulled By: cblappert

fbshipit-source-id: 6696ea54ed9ae7a3d7e8a944c45e14feccf3139b