Don't havoc abstract value args unless really necessary
In #2179 I describe why we might need to havoc some args to abstract values
sometimes. This adds an invariant to ensure that we always have args in
the cases we need to havoc.

If this abstract value cannot be an object, then we don't need to havoc
since there's nothing that can be mutated.

If this has a known values domain, then we know the possible objects. We
only need to havoc those, not the args.