Only havoc abstract value args when really necessary (#2184)
Summary:
Fixes #2179

In #2179 I describe why we might need to havoc some args to abstract values sometimes. This adds an invariant to ensure that we always have args in the cases we need to havoc.

If this abstract value cannot be an object, then we don't need to havoc since there's nothing that can be mutated.

If this has a known values domain, then we know the possible objects. We only need to havoc those, not the args.
Closes https://github.com/facebook/prepack/pull/2184

Differential Revision: D8701601

Pulled By: sebmarkbage

fbshipit-source-id: d5fecbed4cff8013bc768cb6e277094d9ff16d9c