Enforce types of ValuesDomain getElements (#2211)
Summary:
I found this working on another PR but might as well fix it here first.

This wasn't previously enforced and now we are forced to deal with the whether something is a primitive or object as well as some other places we had bugs covered up by `any`.
Closes https://github.com/facebook/prepack/pull/2211

Differential Revision: D8751674

Pulled By: sebmarkbage

fbshipit-source-id: d1445530e53d88a08f3d2652dffd26bcf339d4f3