Make explicit which path conditions apply to a Generator (#2224)
Summary:
Release notes: None

It used to be a case that the Generator constructor simply
harvested whatever was currently set in the realm as the current
path condition.

This makes it explicit, and fixes some places where likely the wrong
path condition was picked up. At least one place caused an issue in
InstantRender that is now fixed (but I couldn't extract a small repro).

Leaving behind a number of `TODO #2222`s in the code to review once more.
Pull Request resolved: https://github.com/facebook/prepack/pull/2224

Differential Revision: D8779258

Pulled By: NTillmann

fbshipit-source-id: a0f3dea4b03eaa71b12943f813eb7fc440d29fca