Small tweak to solve a pressing instance of a more general problem (#2239)
Summary:
Release note: none

Resolves #2238

This is a quick fix for the above issue. The underlying cause is that we do not construct completions with effects, but rely on incremental updates in special situations, such as the constructor for ForkedAbruptCompletions.

I'll try to fix the underlying cause in a subsequent PR.
Pull Request resolved: https://github.com/facebook/prepack/pull/2239

Differential Revision: D8795886

Pulled By: hermanventer

fbshipit-source-id: 677f2481ae3d6463a73f36a7e7106afbd92b068e