Make sure that Completion.effects and Effects.result are consistent (#2244)
Summary:
Release note: none

I'm trying to move from a world where we pass arounds effects and completions separately to a world where only completions are passed around. To get there, we need to get a one to one correspondence between completions and the effects that they complete. This is not currently the case and getting there seems to be quite a bit of work.

To make things more reviewable, I'm breaking out the first part of that. Hence, the todos in the code.

This bit only sets up Effects to always make itself be the value of the effects property of its completion (result). Because of existing cases where a single completion can complete several Effects objects, there is an exemption for such cases, along with todos.
Pull Request resolved: https://github.com/facebook/prepack/pull/2244

Differential Revision: D8825625

Pulled By: hermanventer

fbshipit-source-id: 7fb835d68e806ffd96845983f4d3db6837d1ad5a