Compose subsequent effects with currently applied affects (#2271)
Summary:
Release note: none

This fixes a problem reported as a comment in PR #2258.

The underlying problem was that Realm.evaluateForEffects would just append the normal effects, captured since the last non complete join, into the saved partially normal completion that becomes the result of the returned Effects. This is not quite right because applying the resulting Effects should restore the complete set of normal effects that were in force when evaluateForEffects completed.
Pull Request resolved: https://github.com/facebook/prepack/pull/2271

Differential Revision: D8870763

Pulled By: hermanventer

fbshipit-source-id: 027924f2b250174393c60d90c2feefdb5a55eec3