Move generators from PNC normal branches to joined normal effects (#2274)
Summary:
Release note: none

Generators that end up in joined effects, need to disappear from the generators in the forked completions from which they have been extracted. Also, they need to be applied to the generator containing the timeline in which the fork occurs, before the fork occurs.

With some luck, this may fix a lot of issues.
Pull Request resolved: https://github.com/facebook/prepack/pull/2274

Differential Revision: D8878320

Pulled By: hermanventer

fbshipit-source-id: a6cd41bf9e82bc4e362bda32a1e32c5fc90298cf