Add Basic String Support
This adds basic string support. We'll use UTF8 as the standard format since
it is the most common format coming in and going out. The most common
format to interop with is UTF8 through C, C++, Rust etc. interfaces.

For indexing we can convert to UCS2 in cases we can't prove that it is safe
to index through UTF8.

No ropes will be used to keep implementation lightweight and simple.

Allocation happens on an arena on the stack.
Appending two strings is inlined as memcpy.

Other operations will need a standard library that can be called.