Pretty Print Generators and Effects (#2270)
Summary:
In an effort to make debugging easier, add utilities that help pretty-print generators and effects with all the important fields (and none of the internal ones that you see in a debugger). Depth of recursion of printer is customizable.

Currently doesn't pass Flow because I have a helper function that takes an Object and returns an Object. Not sure how to make that type more specific.

Open to feedback on ways to change the output to be more readable/useful.

Output looks like this for a simple set of effects:
```
{
  result: [ReturnCompletion value [Abstract 1735003607742184]],
  generator: {
    _entries: {
      index: 39,
      declared: [Abstract 1735003607742184],
      args: []
    },
    id: 35,
    _name: evaluateNodeForEffects,
    pathConditions: [[Abstract 1735003607742177], [Abstract 1735003607742183]]
  },
  canBeApplied: true,
  _id: 33
}
```
Pull Request resolved: https://github.com/facebook/prepack/pull/2270

Differential Revision: D8896030

Pulled By: cblappert

fbshipit-source-id: 4f15a047b43d41ccad53222850fad2bf3216b08c