Try conditional branches separately if binary expression fails
If a binary expression can't be resolved as pure (such as when applied
to values of two different types), we can instead try them separately
and join the effects.

If the left value is conditional, we try each possible value separately
until we hit a non-conditional value. Then we try any conditional right
values.