Try conditional branches separately if binary expression fails (#2332)
Summary:
This solves the remaining issues in #2300.

If a binary expression can't be resolved as pure (such as when applied to values of two different types), we can instead try conditions separately and join the effects.

If the left value is conditional, we try each possible value separately until we hit a non-conditional value. Then we try any conditional right values.
Pull Request resolved: https://github.com/facebook/prepack/pull/2332

Differential Revision: D9032487

Pulled By: sebmarkbage

fbshipit-source-id: 0f89a680ad0b710e90c1dd7a4c127c2247527c99