Set on havoced object should target the receiver, not the own object (#2338)
Summary:
In case the havoced object has a setter on it.

This is fixes this particular bug but it's also useful to be able to pass the correct receiver to properties.js rather than unwrapping it like we do now. This will be evident in a follow up PR.

This also lets emitPropertyAssignment deal with abstract values which is a common pattern and will become more common with widened objects.
Pull Request resolved: https://github.com/facebook/prepack/pull/2338

Differential Revision: D9035746

Pulled By: sebmarkbage

fbshipit-source-id: 2abb1a3eb047de1739dec94259a803c4c45e416d